admin/backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==5.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
pymysql==1.1.1
PyJWT==2.10.1
Werkzeug==3.1.3
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4